fix : friend 도메인 패키지 수정
