+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/generic_capsule/repository/CapsuleQueryRepository.java
@@ -56,7 +56,7 @@ public class CapsuleQueryRepository {
             .join(capsule.capsuleSkin, capsuleSkin)
             .join(capsule.member, member)
             .where(ST_Contains(mbr, capsule.point).and(capsule.member.id.eq(memberId)
-                    .and(eqCapsuleType(capsuleType))))
+                .and(eqCapsuleType(capsuleType))))
             .fetch();
     }
