+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/friend/api/command/FriendCommandApi.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.friend.api;
+package site.timecapsulearchive.core.domain.friend.api.command;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -8,61 +8,38 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import java.time.ZonedDateTime;
 import org.springframework.http.ResponseEntity;
-import site.timecapsulearchive.core.domain.friend.data.request.SearchFriendsRequest;
 import site.timecapsulearchive.core.domain.friend.data.request.SendFriendRequest;
-import site.timecapsulearchive.core.domain.friend.data.response.FriendRequestsSliceResponse;
-import site.timecapsulearchive.core.domain.friend.data.response.FriendsSliceResponse;
-import site.timecapsulearchive.core.domain.friend.data.response.SearchFriendsResponse;
-import site.timecapsulearchive.core.domain.friend.data.response.SearchTagFriendSummaryResponse;
 import site.timecapsulearchive.core.global.common.response.ApiSpec;
 import site.timecapsulearchive.core.global.error.ErrorResponse;
 
-public interface FriendApi {
+public interface FriendCommandApi {
 
     @Operation(
-        summary = "소셜 친구 목록 조회",
-        description = "사용자의 소셜 친구 목록을 보여준다.",
+        summary = "소셜 친구 요청 수락",
+        description = "상대방으로부터 온 친구 요청을 수락한다.",
         security = {@SecurityRequirement(name = "user_token")},
         tags = {"friend"}
     )
     @ApiResponses(value = {
         @ApiResponse(
             responseCode = "200",
-            description = "ok"
-        )
-    })
-    ResponseEntity<ApiSpec<FriendsSliceResponse>> findFriends(
-        Long memberId,
-
-        @Parameter(in = ParameterIn.QUERY, description = "페이지 크기", required = true)
-        int size,
-
-        @Parameter(in = ParameterIn.QUERY, description = "마지막 데이터의 시간", required = true)
-        ZonedDateTime createdAt
-    );
-
-    @Operation(
-        summary = "소셜 친구 요청 목록 조회",
-        description = "사용자의 소셜 친구 요청 목록을 보여준다. 수락 대기 중인 요청만 해당한다.",
-        security = {@SecurityRequirement(name = "user_token")},
-        tags = {"friend"}
-    )
-    @ApiResponses(value = {
+            description = "처리 완료"
+        ),
         @ApiResponse(
-            responseCode = "200",
-            description = "ok"
+            responseCode = "404",
+            description = "친구 요청이 존재하지 않는 경우 발생"
+        ),
+        @ApiResponse(
+            responseCode = "500",
+            description = "외부 API 요청 실패"
         )
     })
-    ResponseEntity<ApiSpec<FriendRequestsSliceResponse>> findFriendRequests(
+    ResponseEntity<ApiSpec<String>> acceptFriendRequest(
         Long memberId,
 
-        @Parameter(in = ParameterIn.QUERY, description = "페이지 크기", required = true)
-        int size,
-
-        @Parameter(in = ParameterIn.QUERY, description = "마지막 데이터의 시간", required = true)
-        ZonedDateTime createdAt
+        @Parameter(in = ParameterIn.PATH, required = true, schema = @Schema())
+        Long friendId
     );
 
     @Operation(
@@ -111,7 +88,6 @@ public interface FriendApi {
         Long friendId
     );
 
-
     @Operation(
         summary = "소셜 친구 요청",
         description = "사용자에게 친구 요청을 보낸다. 해당 사용자에게 알림이 발송된다.",
@@ -155,69 +131,5 @@ public interface FriendApi {
         Long memberId,
 
         SendFriendRequest request
-    );
-
-    @Operation(
-        summary = "소셜 친구 요청 수락",
-        description = "상대방으로부터 온 친구 요청을 수락한다.",
-        security = {@SecurityRequirement(name = "user_token")},
-        tags = {"friend"}
-    )
-    @ApiResponses(value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "처리 완료"
-        ),
-        @ApiResponse(
-            responseCode = "404",
-            description = "친구 요청이 존재하지 않는 경우 발생"
-        ),
-        @ApiResponse(
-            responseCode = "500",
-            description = "외부 API 요청 실패"
-        )
-    })
-    ResponseEntity<ApiSpec<String>> acceptFriendRequest(
-        Long memberId,
-
-        @Parameter(in = ParameterIn.PATH, required = true, schema = @Schema())
-        Long friendId
-    );
-
-
-    @Operation(
-        summary = "전화번호를 바탕으로 앱 사용자 조회",
-        description = "전화번호 목록을 받아 ARchive 사용자를 반환한다.",
-        security = {@SecurityRequirement(name = "user_token")},
-        tags = {"friend"}
-    )
-    @ApiResponses(value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "ok"
-        )
-    })
-    ResponseEntity<ApiSpec<SearchFriendsResponse>> searchMembersByPhones(
-        Long memberId,
-        SearchFriendsRequest request
-    );
-
-    @Operation(
-        summary = "찬구 검색",
-        description = "친구의 tag로 친구 검색을 한다.",
-        security = {@SecurityRequirement(name = "user_token")},
-        tags = {"friend"}
-    )
-    @ApiResponses(value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "ok"
-        )
-    })
-    ResponseEntity<ApiSpec<SearchTagFriendSummaryResponse>> searchFriendByTag(
-        Long memberId,
-
-        @Parameter(in = ParameterIn.QUERY, description = "친구 태그", required = true)
-        String tag
     );
 }
