+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/friend/api/query/FriendQueryApiController.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.friend.api;
+package site.timecapsulearchive.core.domain.friend.api.query;
 
 import jakarta.validation.Valid;
 import java.time.ZonedDateTime;
@@ -7,9 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,25 +16,22 @@ import org.springframework.web.bind.annotation.RestController;
 import site.timecapsulearchive.core.domain.friend.data.dto.FriendSummaryDto;
 import site.timecapsulearchive.core.domain.friend.data.dto.SearchFriendSummaryDto;
 import site.timecapsulearchive.core.domain.friend.data.request.SearchFriendsRequest;
-import site.timecapsulearchive.core.domain.friend.data.request.SendFriendRequest;
 import site.timecapsulearchive.core.domain.friend.data.response.FriendRequestsSliceResponse;
 import site.timecapsulearchive.core.domain.friend.data.response.FriendsSliceResponse;
 import site.timecapsulearchive.core.domain.friend.data.response.SearchFriendsResponse;
 import site.timecapsulearchive.core.domain.friend.data.response.SearchTagFriendSummaryResponse;
-import site.timecapsulearchive.core.domain.friend.facade.FriendFacade;
-import site.timecapsulearchive.core.domain.friend.service.FriendService;
+import site.timecapsulearchive.core.domain.friend.service.query.FriendQueryService;
 import site.timecapsulearchive.core.global.common.response.ApiSpec;
 import site.timecapsulearchive.core.global.common.response.SuccessCode;
 import site.timecapsulearchive.core.global.common.wrapper.ByteArrayWrapper;
 import site.timecapsulearchive.core.global.security.encryption.HashEncryptionManager;
 
 @RestController
-@RequestMapping("/friends")
 @RequiredArgsConstructor
-public class FriendApiController implements FriendApi {
+@RequestMapping("/friends")
+public class FriendQueryApiController implements FriendQueryApi {
 
-    private final FriendService friendService;
-    private final FriendFacade friendFacade;
+    private final FriendQueryService friendQueryService;
     private final HashEncryptionManager hashEncryptionManager;
 
     @GetMapping
@@ -46,7 +41,7 @@ public class FriendApiController implements FriendApi {
         @RequestParam(defaultValue = "20", value = "size") final int size,
         @RequestParam(value = "created_at") final ZonedDateTime createdAt
     ) {
-        Slice<FriendSummaryDto> friendsSlice = friendService.findFriendsSlice(memberId, size,
+        Slice<FriendSummaryDto> friendsSlice = friendQueryService.findFriendsSlice(memberId, size,
             createdAt);
 
         return ResponseEntity.ok(
@@ -67,7 +62,7 @@ public class FriendApiController implements FriendApi {
         @RequestParam(defaultValue = "20", value = "size") final int size,
         @RequestParam(value = "created_at") final ZonedDateTime createdAt
     ) {
-        Slice<FriendSummaryDto> friendRequestsSlice = friendService.findFriendRequestsSlice(
+        Slice<FriendSummaryDto> friendRequestsSlice = friendQueryService.findFriendRequestsSlice(
             memberId, size, createdAt);
 
         return ResponseEntity.ok(
@@ -75,81 +70,6 @@ public class FriendApiController implements FriendApi {
                 SuccessCode.SUCCESS,
                 FriendRequestsSliceResponse.createOf(friendRequestsSlice.getContent(),
                     friendRequestsSlice.hasNext())
-            )
-        );
-    }
-
-    @DeleteMapping(value = "/{friend_id}")
-    @Override
-    public ResponseEntity<ApiSpec<String>> deleteFriend(
-        @AuthenticationPrincipal final Long memberId,
-        @PathVariable("friend_id") final Long friendId
-    ) {
-        friendService.deleteFriend(memberId, friendId);
-
-        return ResponseEntity.ok(
-            ApiSpec.empty(
-                SuccessCode.SUCCESS
-            )
-        );
-    }
-
-    @DeleteMapping("{friend_id}/deny-request")
-    @Override
-    public ResponseEntity<ApiSpec<String>> denyFriendRequest(
-        @AuthenticationPrincipal final Long memberId,
-        @PathVariable("friend_id") final Long friendId) {
-
-        friendService.denyRequestFriend(memberId, friendId);
-
-        return ResponseEntity.ok(
-            ApiSpec.empty(
-                SuccessCode.SUCCESS
-            )
-        );
-    }
-
-    @PostMapping(value = "/{friend_id}/request")
-    @Override
-    public ResponseEntity<ApiSpec<String>> requestFriend(
-        @AuthenticationPrincipal final Long memberId,
-        @PathVariable("friend_id") final Long friendId) {
-
-        friendService.requestFriend(memberId, friendId);
-
-        return ResponseEntity.ok(
-            ApiSpec.empty(
-                SuccessCode.ACCEPTED
-            )
-        );
-    }
-
-    @PostMapping(value = "/requests")
-    @Override
-    public ResponseEntity<ApiSpec<String>> requestFriends(
-        @AuthenticationPrincipal final Long memberId,
-        @RequestBody SendFriendRequest request
-    ) {
-        friendFacade.requestFriends(memberId, request.friendIds());
-
-        return ResponseEntity.ok(
-            ApiSpec.empty(
-                SuccessCode.ACCEPTED
-            )
-        );
-    }
-
-    @PostMapping(value = "/{friend_id}/accept-request")
-    @Override
-    public ResponseEntity<ApiSpec<String>> acceptFriendRequest(
-        @AuthenticationPrincipal final Long memberId,
-        @PathVariable("friend_id") final Long friendId
-    ) {
-        friendService.acceptFriend(memberId, friendId);
-
-        return ResponseEntity.ok(
-            ApiSpec.empty(
-                SuccessCode.SUCCESS
             )
         );
     }
@@ -167,7 +87,7 @@ public class FriendApiController implements FriendApi {
         final List<ByteArrayWrapper> phoneEncryption = request.toPhoneEncryption(
             hashEncryptionManager::encrypt);
 
-        final List<SearchFriendSummaryDto> dtos = friendService.findFriendsByPhone(
+        final List<SearchFriendSummaryDto> dtos = friendQueryService.findFriendsByPhone(
             memberId, phoneEncryption);
 
         return ResponseEntity.ok(
@@ -187,7 +107,7 @@ public class FriendApiController implements FriendApi {
         return ResponseEntity.ok(
             ApiSpec.success(
                 SuccessCode.SUCCESS,
-                friendService.searchFriend(memberId, tag)
+                friendQueryService.searchFriend(memberId, tag)
             )
         );
     }
