+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/friend/repository/friend_invite/FriendInviteQueryRepository.java
@@ -1,0 +1,8 @@
+package site.timecapsulearchive.core.domain.friend.repository.friend_invite;
+
+import java.util.List;
+
+public interface FriendInviteQueryRepository {
+
+    void bulkSave(final Long ownerId, final List<Long> friendIds);
+}
