+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/friend/repository/member_friend/MemberFriendQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.friend.repository;
+package site.timecapsulearchive.core.domain.friend.repository.member_friend;
 
 import static site.timecapsulearchive.core.domain.friend.entity.QFriendInvite.friendInvite;
 import static site.timecapsulearchive.core.domain.friend.entity.QMemberFriend.memberFriend;
@@ -22,7 +22,7 @@ import site.timecapsulearchive.core.global.common.wrapper.ByteArrayWrapper;
 
 @Repository
 @RequiredArgsConstructor
-public class MemberFriendQueryRepository {
+public class MemberFriendQueryRepositoryImpl implements MemberFriendQueryRepository {
 
     private final JPAQueryFactory jpaQueryFactory;
