+++ backend/core/src/test/java/site/timecapsulearchive/core/common/dependency/UnitTestDependency.java
@@ -8,7 +8,6 @@ import site.timecapsulearchive.core.global.security.encryption.HashProperties;
 import site.timecapsulearchive.core.infra.s3.config.S3Config;
 import site.timecapsulearchive.core.infra.s3.config.S3Properties;
 import site.timecapsulearchive.core.infra.s3.manager.S3PreSignedUrlManager;
-import site.timecapsulearchive.core.infra.s3.manager.S3UrlGenerator;
 
 /**
  * 유닛 테스트를 위한 의존성 클래스 집합체
