+++ backend/core/src/test/java/site/timecapsulearchive/core/domain/capsule/public_capsule/service/PublicCapsuleServiceTest.java
@@ -11,7 +11,6 @@ import org.springframework.context.annotation.Import;
 import site.timecapsulearchive.core.common.fixture.dto.CapsuleDtoFixture;
 import site.timecapsulearchive.core.domain.capsule.generic_capsule.data.dto.CapsuleDetailDto;
 import site.timecapsulearchive.core.domain.capsule.public_capsule.repository.PublicCapsuleQueryRepository;
-import site.timecapsulearchive.core.domain.capsule.public_capsule.service.PublicCapsuleService;
 import site.timecapsulearchive.core.global.geography.GeoTransformConfig;
 import site.timecapsulearchive.core.infra.s3.config.S3Config;
